+++ kasper-api/src/main/java/com/viadeo/kasper/ddd/annotation/XKasperDomain.java
@@ -34,5 +34,10 @@ public @interface XKasperDomain {
 	 * @return the domain's description
 	 */
 	String description() default "";
+
+ 	/**
+	 * @return the domain's owner (suggested syntax: 'Firstname Lastname <user@domain.dtd>')
+	 */
+	String owner() default "";
 	
 }