+++ kasper-common/src/main/java/com/viadeo/kasper/tools/ObjectMapperProvider.java
@@ -34,7 +34,7 @@ public final class ObjectMapperProvider {
     static final String ERRORS = "errors";
     static final String MESSAGE = "message";
     static final String CODE = "code";
-    static final String USERMESSAGE = "userMessage";
+    static final String LABEL = "label";
     static final String STATUS = "status";
 
     public static final ObjectMapperProvider INSTANCE = new ObjectMapperProvider();
@@ -43,7 +43,7 @@ public final class ObjectMapperProvider {
 
     // ------------------------------------------------------------------------
 
-    private ObjectMapperProvider() {
+    public ObjectMapperProvider() {
         mapper = new ObjectMapper();
 
         /* Generic features */
@@ -76,6 +76,7 @@ public final class ObjectMapperProvider {
         kasperClientModule.setDeserializers(new CommandQueryResponseDeserializerAdapter());
 
         mapper.registerModule(kasperClientModule);
+        mapper.registerModule(new KasperImmutabilityParanamerModule());
 
         /* Third-party modules */
         mapper.registerModule(new GuavaModule());