+++ kasper-platform/src/test/java/com/viadeo/kasper/core/metrics/KasperMetricsITest.java
@@ -10,7 +10,8 @@ import com.codahale.metrics.MetricRegistry;
 import com.viadeo.kasper.client.platform.Platform;
 import com.viadeo.kasper.client.platform.configuration.KasperPlatformConfiguration;
 import com.viadeo.kasper.client.platform.domain.sample.MyCustomDomainBox;
-import com.viadeo.kasper.core.resolvers.CommandResolver;
+import com.viadeo.kasper.context.Context;
+import com.viadeo.kasper.context.impl.DefaultContextBuilder;
 import com.viadeo.kasper.event.EventListener;
 import com.viadeo.kasper.exception.KasperException;
 import org.junit.Test;
@@ -29,7 +30,6 @@ public class KasperMetricsITest {
 
         // clear caches in order to ensure test integrity
         KasperMetrics.clearCache();
-        new CommandResolver().clearCache();
     }
 
     // ------------------------------------------------------------------------
@@ -132,6 +132,50 @@ public class KasperMetricsITest {
     }
 
     @Test
+    public void name_fromMyCustomCommand_withDomainTypeAsStyle_shouldBeOk() {
+        // Given
+        buildPlatformWith(new MetricRegistry());
+
+        // When
+        final String name = KasperMetrics.name(
+                MetricNameStyle.DOMAIN_TYPE,
+                MyCustomDomainBox.MyCustomCommand.class,
+                "bip"
+        );
+
+        // Then
+        assertEquals("mycustomdomain.command.bip", name);
+    }
+
+    @Test
+    public void name_fromMyCustomCommand_withDomainTypeComponentAsStyle_shouldBeOk() {
+        // Given
+        buildPlatformWith(new MetricRegistry());
+
+        // When
+        final String name = KasperMetrics.name(
+                MetricNameStyle.DOMAIN_TYPE_COMPONENT,
+                MyCustomDomainBox.MyCustomCommand.class,
+                "bip"
+        );
+
+        // Then
+        assertEquals("mycustomdomain.command.mycustomcommand.bip", name);
+    }
+
+    @Test
+    public void name_fromMyCustomCommand_withNoneAsStyle_shouldBeOk() {
+        // Given
+        buildPlatformWith(new MetricRegistry());
+
+        // When
+        final String name = KasperMetrics.name(MetricNameStyle.NONE, MyCustomDomainBox.MyCustomCommand.class, "bip");
+
+        // Then
+        assertEquals("com.viadeo.kasper.client.platform.domain.sample.mycustomdomainbox$mycustomcommand.bip", name);
+    }
+
+    @Test
     public void name_fromMyCustomQuery_shouldBeOk() {
         // Given
         buildPlatformWith(new MetricRegistry());
@@ -188,5 +232,63 @@ public class KasperMetricsITest {
         KasperMetrics.name(MyCustomDomainBox.MyCustomMalformedDomainEvent.class, "bip");
 
         // Then throws an exception
+    }
+
+    @Test
+    public void name_onClientPerType_fromMyCustomQuery_withSpecifiedApplicationId_shouldBeOk() {
+        // Given
+        buildPlatformWith(new MetricRegistry());
+
+        final Context context = DefaultContextBuilder.get();
+        context.setApplicationId("foobar");
+
+        // When
+        final String name = KasperMetrics.name(MetricNameStyle.CLIENT_TYPE, context, MyCustomDomainBox.MyCustomQuery.class, "bip");
+
+        // Then
+        assertEquals("client.foobar.query.bip", name);
+    }
+
+    @Test
+    public void name_onClientPerType_fromMyCustomQuery_withUnspecifiedApplicationId_shouldBeOk() {
+        // Given
+        buildPlatformWith(new MetricRegistry());
+
+        final Context context = DefaultContextBuilder.get();
+
+        // When
+        final String name = KasperMetrics.name(MetricNameStyle.CLIENT_TYPE, context, MyCustomDomainBox.MyCustomQuery.class, "bip");
+
+        // Then
+        assertEquals("client.unknown.query.bip", name);
+    }
+
+    @Test
+    public void name_onClientPerType_fromMyCustomCommand_withSpecifiedApplicationId_shouldBeOk() {
+        // Given
+        buildPlatformWith(new MetricRegistry());
+
+        final Context context = DefaultContextBuilder.get();
+        context.setApplicationId("foobar");
+
+        // When
+        final String name = KasperMetrics.name(MetricNameStyle.CLIENT_TYPE, context, MyCustomDomainBox.MyCustomCommand.class, "bip");
+
+        // Then
+        assertEquals("client.foobar.command.bip", name);
+    }
+
+    @Test
+    public void name_onClientPerType_fromMyCustomCommand_withUnspecifiedApplicationId_shouldBeOk() {
+        // Given
+        buildPlatformWith(new MetricRegistry());
+
+        final Context context = DefaultContextBuilder.get();
+
+        // When
+        final String name = KasperMetrics.name(MetricNameStyle.CLIENT_TYPE, context, MyCustomDomainBox.MyCustomCommand.class, "bip");
+
+        // Then
+        assertEquals("client.unknown.command.bip", name);
     }
 }