+++ kasper-test/src/main/java/com/viadeo/kasper/test/platform/KasperPlatformCommandResultValidator.java
@@ -15,8 +15,8 @@ import org.axonframework.test.AxonAssertionError;
 import java.util.Arrays;
 import java.util.Iterator;
 
-import static com.viadeo.kasper.cqrs.command.CommandResponse.Status.*;
-import static com.viadeo.kasper.test.matchers.KasperMatcher.equalTo;
+import static com.viadeo.kasper.KasperResponse.Status.*;
+import static com.viadeo.kasper.tools.KasperMatcher.equalTo;
 
 /**
  * FIXME: add better debugging information