+++ kasper-test/src/test/java/com/viadeo/kasper/cqrs/command/TestFixtureAxonTest.java
@@ -29,7 +29,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static com.viadeo.kasper.cqrs.command.FixtureUseCase.*;
-import static com.viadeo.kasper.test.matchers.KasperMatcher.equalTo;
+import static com.viadeo.kasper.tools.KasperMatcher.equalTo;
 import static org.axonframework.test.matchers.Matchers.*;
 import static org.mockito.Mockito.mock;